Group dependency injection modules together for easier mocking in tests
