--- app/src/main/java/ch/epfl/cs311/wanderwave/di/AppModule.kt
@@ -1,7 +1,0 @@
-package ch.epfl.cs311.wanderwave.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module @InstallIn(SingletonComponent::class) object AppModule
