--- app/src/main/java/ch/epfl/cs311/wanderwave/di/OkHttpModule.kt
@@ -1,19 +1,0 @@
-package ch.epfl.cs311.wanderwave.di
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-import okhttp3.OkHttpClient
-
-@Module
-@InstallIn(SingletonComponent::class)
-object OkHttpModule {
-
-  @Provides
-  @Singleton
-  fun provideOkHttpClient(): OkHttpClient {
-    return OkHttpClient()
-  }
-}
