--- app/src/main/java/ch/epfl/cs311/wanderwave/di/SpotifyModule.kt
@@ -1,21 +1,0 @@
-package ch.epfl.cs311.wanderwave.di
-
-import android.content.Context
-import ch.epfl.cs311.wanderwave.model.spotify.SpotifyController
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object SpotifyModule {
-
-  @Provides
-  @Singleton
-  fun provideSpotifyController(@ApplicationContext context: Context): SpotifyController {
-    return SpotifyController(context)
-  }
-}
